Refactored mobile dtos
