+++ server/models/http/responses/mobile_comment_response_models.py
@@ -3,21 +3,10 @@ from pydantic import BaseModel
 
 from server.models.database.mobile_comments_db_model import Comment
 
-
-class MobileCommentRegister(BaseModel):
+class MobileCommentResponse(BaseModel):
     comment: str
     email: str | None
-    
-
-def to_comment_model(c: MobileCommentRegister) -> Comment:
-    return Comment(
-        comment=c.comment,
-        email=c.email
-    )
-
-class MobileCommentResponse(MobileCommentRegister):
     created_at: datetime
-
     
     @classmethod
     def from_model(c, comment: Comment):
