+++ server/routes/public/mobile_classes_routes.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter
 from server.deps.session_dep import SessionDep
 from server.models.database.class_db_model import Class
 from server.repositories.class_repository import ClassRepository
-from server.routes.public.dtos.mobile_class_response import MobileClassResponse
+from server.models.http.responses.mobile_class_response_models import MobileClassResponse
 
 router = APIRouter(prefix="/mobile/classes", tags=["Mobile","Classes"])
