+++ server/routes/public/mobile_programs_routes.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+from server.deps.session_dep import SessionDep
+
+router = APIRouter(prefix="/mobile/programs", tags=["Mobile","Classes"])
+
+@router.get("")
+async def get_all_programs(session: SessionDep) -> list:
+    """Get all programs (?)"""
+    return []
